+++ ClickHouse.Driver.Tests/Logging/ClickHouseLoggingTests.cs
@@ -1,8 +1,8 @@
 #if NET7_0_OR_GREATER
 
 using System.Net.Http;
+using System.Threading.Tasks;
 using ClickHouse.Driver.ADO;
-using ClickHouse.Driver.Utility;
 
 namespace ClickHouse.Driver.Tests.Logging;
 
@@ -17,7 +17,7 @@ public class ClickHouseLoggingTests
         {
             LoggerFactory = factory,
         };
-        
+
         try
         {
             using var connection = dataSource.CreateConnection();
@@ -27,6 +27,53 @@ public class ClickHouseLoggingTests
         {
             dataSource.Dispose();
         }
+    }
+
+    [Test]
+    public async Task Connection_WithEnableDebugMode_ActivatesTraceHelper()
+    {
+        // Arrange
+        var factory = new CapturingLoggerFactory();
+        factory.MinimumLevel = Microsoft.Extensions.Logging.LogLevel.Trace;
+        
+        var builder = TestUtilities.GetConnectionStringBuilder();
+        var settings = new ClickHouseClientSettings(builder)
+        {
+            LoggerFactory = factory,
+            EnableDebugMode = true,
+        };
+
+        // Act
+        using var connection = new ClickHouseConnection(settings);
+        
+        await connection.OpenAsync();
+
+        // Assert - TraceHelper should create a logger for NetTrace when activated
+        Assert.That(factory.Loggers, Does.ContainKey("ClickHouse.Driver.NetTrace"),
+            "EnableDebugMode with LoggerFactory should activate TraceHelper and create NetTrace logger");
+    }
+
+    [Test]
+    public async Task Connection_WithoutEnableDebugMode_DoesNotActivateTraceHelper()
+    {
+        // Arrange
+        var factory = new CapturingLoggerFactory();
+        
+        var builder = TestUtilities.GetConnectionStringBuilder();
+        var settings = new ClickHouseClientSettings(builder)
+        {
+            LoggerFactory = factory,
+            EnableDebugMode = false,
+        };
+
+        // Act
+        using var connection = new ClickHouseConnection(settings);
+        
+        await connection.OpenAsync();
+
+        // Assert - TraceHelper should not create a NetTrace logger when not enabled
+        Assert.That(factory.Loggers, Does.Not.ContainKey("ClickHouse.Driver.NetTrace"),
+            "Without EnableDebugMode, TraceHelper should not be activated");
     }
 }
 #endif